Extract Prompt Studio widget helpers
