+++ docs/development/issue-14-frontend-js-maintenance.md
@@ -209,6 +209,7 @@ web/js/
     dom.js
     style.js
     tooltip.js
+    widgets.js
     fields.js
     node_hooks.js
 ```
@@ -232,9 +233,10 @@ Planned slices:
 9. `wheel.js`
 10. `dom.js` and `style.js`
 11. `tooltip.js`
-12. `fields.js`
-13. `node_hooks.js`
-14. entry file slim-down
+12. `widgets.js`
+13. `fields.js`
+14. `node_hooks.js`
+15. entry file slim-down
 
 Each PR should mostly move code and update imports. Behavior changes should be
 separate PRs unless the move reveals a confirmed bug that cannot be separated.
@@ -713,6 +715,7 @@ Use this in Issue #14 or follow-up PR tracking comments.
 [ ] dom.js extracted
 [ ] style.js extracted
 [ ] tooltip.js extracted
+[ ] widgets.js extracted
 [ ] fields.js extracted
 [ ] node_hooks.js extracted
 [ ] entry file slimmed down
